+++ BitMono/BitMono.Obfuscation.API/IDependenciesDataResolver.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace BitMono.Obfuscation.API
+{
+    public interface IDependenciesDataResolver
+    {
+        IEnumerable<byte[]> Resolve();
+    }
+}